1.0.12 - 15.01.2026 18:29

+++ src/main/java/com/roften/avilixeconomy/client/ClientBalance.java
@@ -1,8 +1,8 @@
 package com.roften.avilixeconomy.client;
 
 public class ClientBalance {
-    private static long balance = 0;
+    private static double balance = 0.0;
 
-    public static long get() { return balance; }
-    public static void set(long v) { balance = v; }
+    public static double get() { return balance; }
+    public static void set(double v) { balance = v; }
 }
